Added functions rcm2rgrid and rgrid2rcm. Made a parameter in filwgts_lancos a numeric.
git-svn-id: https://subversion.ucar.edu/ncldev/ncarg/trunk@6384 10fe2890-8a40-4696-8b1d-7ccbb19c398e

+++ ni/src/lib/nfp/filterxW.c
@@ -14,7 +14,7 @@
 #include <ncarg/ncl/NclCallBacksI.h>
 #include <math.h>
 
-extern void NGCALLF(dfiltrq,DFILTRQ)(int*,double*,double*,int*,int*,
+extern void NGCALLF(dfiltrq,DFILTRQ)(int*,double*,double*,double*,int*,
                                      double*,double*,double*,int*);
 
 NhlErrorTypes filwgts_lancos_W( void )
